refactor: isolate image SSE parsing
